add reset button
